ETH2 packages types fix (#3878)
* Wrap function params in object

* Wrap function params in object

* Export needed typings in web3-eth2-core. Export class definition in web3-eth2-beaconchain

* Replace interface type with class

* Remove addBlockExplorerApi from class type

* linter errors. Add ETH2Core class export

* Linter errors

* Update changelog

* Replace index.d.ts files with types.ts

* Update type declaration files, update dependencies for web3-eth2-core

* remove dtslint commands

* Add dtslint npm command - I think lerna needs it to report a success

* Remove dtslint command

* dtslint debug

* dtslint debug

* Remove dtslint from devDependencies and remove dtslint npm command

* Remove dtslint with npm remove

* Update types file path

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* Debug

* tsc errors

* Remove types property

* Debug tsconfig

* Add .babelrc

* Debug

* Update with PR numbers
