--- packages/web3-eth2-beaconchain/types/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "lib": ["es6"],
-        "target": "es6",
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "strictNullChecks": true,
-        "strictFunctionTypes": true,
-        "noEmit": true,
-        "allowSyntheticDefaultImports": false,
-        "baseUrl": ".",
-        "paths": {
-            "web3-eth2-beaconchain": ["."]
-        }
-    }
-}
