+++ packages/web3-eth2-core/package.json
@@ -7,20 +7,18 @@
     "engines": {
         "node": ">=8.0.0"
     },
-    "types": "types/index.d.ts",
+    "types": "types.ts",
     "scripts": {
-        "dtslint": "dtslint --localTs ../../node_modules/typescript/lib types",
         "test": "jest",
-        "tsc": "tsc -b tsconfig.json"
+        "tsc": "babel src --out-dir lib --extensions \".ts,.tsx\" --source-maps inline"
     },
     "main": "lib/index.js",
     "dependencies": {
-        "axios": "^0.20.0"
+        "axios": "^0.21.1"
     },
     "devDependencies": {
         "@types/jest": "^26.0.14",
-        "@types/node": "^14.11.8",
-        "dtslint": "^3.4.1",
+        "@types/node": "^14.14.22",
         "jest": "^26.5.3",
         "ts-jest": "^26.4.1",
         "typescript": "^3.9.5"
