+++ packages/web3-eth2-core/tsconfig.json
@@ -1,9 +1,21 @@
 {
-  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "./lib"
-  },
-  "include": [
-    "src"
-  ]
- }
+    "declaration": true,
+    "outDir": "lib",
+    "strict": true,
+    "esModuleInterop": true,
+    "module": "commonjs",
+    "lib": ["es6"],
+    "target": "es6",
+    "noImplicitAny": true,
+    "noImplicitThis": true,
+    "strictNullChecks": true,
+    "strictFunctionTypes": true,
+    "noEmit": true,
+    "allowSyntheticDefaultImports": false,
+    "baseUrl": ".",
+    "paths": {
+        "web3-eth2-core": ["."]
+    }
+  }
+}
