+++ packages/web3-eth2-core/tslint.json
@@ -7,6 +7,7 @@
         "whitespace": false,
         "no-unnecessary-class": false,
         "unified-signatures": false,
-        "no-redundant-jsdoc": false
+        "no-redundant-jsdoc": false,
+        "interface-name": [true, "always-prefix"]
     }
 }
