--- packages/web3-eth2-core/types/index.d.ts
@@ -1,5 +1,0 @@
-export interface ETH2BaseOpts {
-    protectProvider?: boolean
-}
-
-export type ETH2Function<T = any> = (...args: any[]) => Promise<T>;
